build(lipwig): renamed lipwig-nest to lipwig
