+++ angular.json
@@ -10,9 +10,9 @@
     "fps": "libs/fps",
     "french-toast": "apps/french-toast",
     "jabberwocky": "apps/jabberwocky",
+    "lipwig": "apps/lipwig",
     "lipwig-admin": "apps/lipwig-admin",
     "lipwig-js": "libs/lipwig-js",
-    "lipwig-nest": "apps/lipwig-nest",
     "lipwig-ui": "libs/lipwig-ui",
     "ridcully": "apps/ridcully",
     "ships": "apps/ships",
