+++ apps/lipwig/project.json
@@ -1,7 +1,7 @@
 {
-  "name": "lipwig-nest",
+  "name": "lipwig",
   "$schema": "../../node_modules/nx/schemas/project-schema.json",
-  "sourceRoot": "apps/lipwig-nest/src",
+  "sourceRoot": "apps/lipwig/src",
   "projectType": "application",
   "targets": {
     "build": {
@@ -10,10 +10,10 @@
       "options": {
         "target": "node",
         "compiler": "tsc",
-        "outputPath": "dist/apps/lipwig-nest",
-        "main": "apps/lipwig-nest/src/main.ts",
-        "tsConfig": "apps/lipwig-nest/tsconfig.app.json",
-        "assets": ["apps/lipwig-nest/src/assets"]
+        "outputPath": "dist/apps/lipwig",
+        "main": "apps/lipwig/src/main.ts",
+        "tsConfig": "apps/lipwig/tsconfig.app.json",
+        "assets": ["apps/lipwig/src/assets"]
       },
       "configurations": {
         "production": {
@@ -22,8 +22,8 @@
           "inspect": false,
           "fileReplacements": [
             {
-              "replace": "apps/lipwig-nest/src/environments/environment.ts",
-              "with": "apps/lipwig-nest/src/environments/environment.prod.ts"
+              "replace": "apps/lipwig/src/environments/environment.ts",
+              "with": "apps/lipwig/src/environments/environment.prod.ts"
             }
           ]
         }
@@ -32,11 +32,11 @@
     "serve": {
       "executor": "@nrwl/js:node",
       "options": {
-        "buildTarget": "lipwig-nest:build"
+        "buildTarget": "lipwig:build"
       },
       "configurations": {
         "production": {
-          "buildTarget": "lipwig-nest:build:production"
+          "buildTarget": "lipwig:build:production"
         }
       }
     },
@@ -44,14 +44,14 @@
       "executor": "@nrwl/linter:eslint",
       "outputs": ["{options.outputFile}"],
       "options": {
-        "lintFilePatterns": ["apps/lipwig-nest/**/*.ts"]
+        "lintFilePatterns": ["apps/lipwig/**/*.ts"]
       }
     },
     "test": {
       "executor": "@nrwl/jest:jest",
       "outputs": ["{workspaceRoot}/coverage/{projectRoot}"],
       "options": {
-        "jestConfig": "apps/lipwig-nest/jest.config.ts",
+        "jestConfig": "apps/lipwig/jest.config.ts",
         "passWithNoTests": true
       }
     }
