+++ apps/lipwig/src/app/lipwig.model.ts
@@ -1,7 +1,7 @@
 import type { WebSocket } from 'ws';
 
 export interface LipwigSocket extends WebSocket {
-    id: string;
-    room: string;
-    host: boolean;
+  id: string;
+  room: string;
+  host: boolean;
 }
